backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
groq>=0.12.0
pymongo>=4.10.1
python-dotenv>=1.0.1
pydantic>=2.10.3
